requirements.txt
google-generativeai
watchdog
fastapi
uvicorn
python-dotenv
pyyaml
sqlalchemy
requests
